Cleanup: Remove unused files.
[#171656112]
https://www.pivotaltracker.com/story/show/171656112
